requirements.txt
Flask-API.yandex
flask
marshmallow
aniso8601==3.0.2
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.16.0
python-dotenv==0.9.1
pytz==2018.5
six==1.11.0
Werkzeug==0.14.1
